+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 1.0.0 under development
 
 - Initial release.
+- Enh #300: Collect call location for `push()` and `status()` in `QueueCollector` (WarLikeLaux)